requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
asgiref==3.7.2
attrs==23.1.0
discord.py==2.3.2
Django==4.2.7
frozenlist==1.4.0
idna==3.6
image==1.5.33
multidict==6.0.4
mysql-connector-python==8.2.0
packaging==23.2
Pillow==10.1.0
protobuf==4.21.12
six==1.16.0
sqlparse==0.4.4
websocket-client==1.6.4
yarl==1.9.3
